api/chatapp/embeddings/build_embeddings.py
# Ejemplo dummy
# Aquí implementarías la lógica para:
# 1. Leer el Cornell Movie Dialogs Corpus
# 2. Generar embeddings con SentenceTransformers u otro modelo
# 3. Crear índice FAISS
# 4. Guardar el índice en faiss_index.bin
#
# Ejemplo (no funcional):
import faiss
import numpy as np

# Dimensión de embeddings
d = 768
index = faiss.IndexFlatL2(d)

# Generar vectores aleatorios
data = np.random.rand(1000, d).astype('float32')
index.add(data)

faiss.write_index(index, "faiss_index.bin")
