api/chatapp/requirements.txt
Django>=3.2
djangorestframework
transformers
django-cors-headers
torch
faiss-cpu
requests
